// src/types.ts
export enum ImageFormats {
    PNG = 'png',
    JPEG = 'jpeg',
    SVG = 'svg',
}

export enum ScreenshotType {
    Browser = 'Browser',
    Device = 'Device',
    None = 'None',
    Twitter = 'Twitter',
    Code = 'Code'
}

export enum CanvasBackgroundTypes {
    Image = 'Image',
    Solid = 'Solid',
    Gradient = 'Gradient',
    None = 'None',
}

export enum Browsers {
    Edge,
    Chrome,
    Firefox,
    Safari,
}

export enum SocialProviders {
    Facebook,
    Twitter,
}